Split `MessageDeliveryService` into two, use `Inbox` operations asynchronously. Use separate executors for client's gRPC routines and Spine's shard observer. Remove excessive logging on the client side.

+++ client/demo/src/main/java/io/spine/message/delivery/demo/GreetArmyServlet.java
@@ -96,8 +96,8 @@ public final class GreetArmyServlet extends ContextAwareServlet {
     private void logPerformance(int howManySoldiers, long startMillis) {
         long endMillis = System.currentTimeMillis();
         long durationMillis = endMillis - startMillis;
-        long perSoldier = durationMillis / howManySoldiers;
-        _info().log("Army was greeted for %d ms. That's about %d ms per soldier.",
+        double perSoldier = durationMillis / (double) howManySoldiers;
+        _info().log("Army was greeted for %d ms. That's about %f ms per soldier.",
                     durationMillis, perSoldier);
     }
 }
